Dropping work
Shooter bases updated, spin no longer working, rotation broken uncertain to cause.
