--- Assets/Scripts/Enemy/ShooterArc.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 907cbefd03eb843418a3bce21709a605
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
